requirements.txt
appdirs==1.4.3
astroid==1.5.2
autopep8==1.3.1
autopy==0.51
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
dodgy==0.1.9
enum34==1.1.6
flake8==3.3.0
functools32==3.2.3.post2
glog==0.3.1
ipython==5.3.0
ipython-genutils==0.2.0
isort==4.2.5
lazy-object-proxy==1.3.1
matplotlib==2.0.1
mccabe==0.6.1
mss==2.0.22
networkx==1.11
numpy==1.12.1
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pep8-naming==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
prospector==0.12.5
psutil==5.2.2
ptyprocess==0.5.1
pudb==2017.1.2
py==1.4.33
PyAutoGUI==0.9.36
pycodestyle==2.3.1
pydocstyle==1.0.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.1
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
PyMsgBox==1.0.6
pyparsing==2.2.0
PyScreeze==0.1.9
pytest==3.0.7
python-dateutil==2.6.0
python-gflags==3.1.1
python-xlib==0.19
PyTweening==1.0.3
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
requests==2.13.0
requirements-detector==0.5.2
scandir==1.5
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
setoptconf==0.2.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
subprocess32==3.2.7
traitlets==4.3.2
urwid==1.3.1
wcwidth==0.1.7
wrapt==1.10.10
yapf==0.16.1
python-xlib
